refactor(27085): restructure the publish toolbar
